Modularise unary ops (#3575)
Modularize abs, acos, acosh, asin, asinh, atan, atanh, cos, cosh, sin, sinh, tan, tanh.

DEV
