+++ tfjs-core/src/ops/sin_test.ts
@@ -1,0 +1,111 @@
+/**
+ * @license
+ * Copyright 2020 Google LLC. All Rights Reserved.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * =============================================================================
+ */
+
+import * as tf from '../index';
+import {ALL_ENVS, describeWithFlags} from '../jasmine_util';
+import {expectArraysClose} from '../test_util';
+
+describeWithFlags('sin', ALL_ENVS, () => {
+  it('basic', async () => {
+    const values = [1, -3, 2, 7, -4];
+    const a = tf.tensor1d(values);
+    const result = tf.sin(a);
+
+    const expected = [];
+    for (let i = 0; i < a.size; i++) {
+      expected[i] = Math.sin(values[i]);
+    }
+    expectArraysClose(await result.data(), expected);
+  });
+
+  it('propagates NaNs', async () => {
+    const a = tf.tensor1d([4, NaN, 0]);
+    const res = tf.sin(a);
+    expectArraysClose(await res.data(), [Math.sin(4), NaN, Math.sin(0)]);
+  });
+
+  it('gradients: Scalar', async () => {
+    const a = tf.scalar(5);
+    const dy = tf.scalar(8);
+
+    const gradients = tf.grad(a => tf.sin(a))(a, dy);
+
+    expect(gradients.shape).toEqual(a.shape);
+    expect(gradients.dtype).toEqual('float32');
+    expectArraysClose(await gradients.data(), [8 * Math.cos(5)]);
+  });
+
+  it('gradient with clones', async () => {
+    const a = tf.scalar(5);
+    const dy = tf.scalar(8);
+
+    const gradients = tf.grad(a => tf.sin(a.clone()).clone())(a, dy);
+
+    expect(gradients.shape).toEqual(a.shape);
+    expect(gradients.dtype).toEqual('float32');
+    expectArraysClose(await gradients.data(), [8 * Math.cos(5)]);
+  });
+
+  it('gradients: Tensor1D', async () => {
+    const a = tf.tensor1d([-1, 2, 3, -5]);
+    const dy = tf.tensor1d([1, 2, 3, 4]);
+
+    const gradients = tf.grad(a => tf.sin(a))(a, dy);
+
+    expect(gradients.shape).toEqual(a.shape);
+    expect(gradients.dtype).toEqual('float32');
+    expectArraysClose(
+        await gradients.data(),
+        [1 * Math.cos(-1), 2 * Math.cos(2), 3 * Math.cos(3), 4 * Math.cos(-5)],
+        1e-1);
+  });
+
+  it('gradients: Tensor2D', async () => {
+    const a = tf.tensor2d([-3, 1, 2, 3], [2, 2]);
+    const dy = tf.tensor2d([1, 2, 3, 4], [2, 2]);
+
+    const gradients = tf.grad(a => tf.sin(a))(a, dy);
+
+    expect(gradients.shape).toEqual(a.shape);
+    expect(gradients.dtype).toEqual('float32');
+    expectArraysClose(
+        await gradients.data(),
+        [1 * Math.cos(-3), 2 * Math.cos(1), 3 * Math.cos(2), 4 * Math.cos(3)],
+        1e-1);
+  });
+
+  it('throws when passed a non-tensor', () => {
+    expect(() => tf.sin({} as tf.Tensor))
+        .toThrowError(/Argument 'x' passed to 'sin' must be a Tensor/);
+  });
+
+  it('accepts a tensor-like object', async () => {
+    const values = [1, -3, 2, 7, -4];
+    const result = tf.sin(values);
+
+    const expected = [];
+    for (let i = 0; i < values.length; i++) {
+      expected[i] = Math.sin(values[i]);
+    }
+    expectArraysClose(await result.data(), expected);
+  });
+
+  it('throws for string tensor', () => {
+    expect(() => tf.sin('q'))
+        .toThrowError(/Argument 'x' passed to 'sin' must be numeric/);
+  });
+});
