+++ tfjs-core/src/ops/unary_ops_test.ts
@@ -17,155 +17,7 @@
 
 import * as tf from '../index';
 import {ALL_ENVS, describeWithFlags} from '../jasmine_util';
-import {expectArraysClose, TEST_EPSILON_FLOAT16} from '../test_util';
-import * as util from '../util';
-
-describeWithFlags('abs', ALL_ENVS, () => {
-  it('basic', async () => {
-    const a = tf.tensor1d([1, -2, 0, 3, -0.1]);
-    const result = tf.abs(a);
-    expectArraysClose(await result.data(), [1, 2, 0, 3, 0.1]);
-  });
-
-  it('5D', async () => {
-    const a = tf.tensor5d([1, -2, 0, -3], [1, 2, 2, 1, 1]);
-    const result = tf.abs(a);
-    expectArraysClose(await result.data(), [1, 2, 0, 3]);
-  });
-
-  it('6D', async () => {
-    const a = tf.tensor6d([1, -2, 5, -3, -1, 4, 7, 8], [1, 2, 2, 2, 1, 1]);
-    const result = tf.abs(a);
-    expectArraysClose(await result.data(), [1, 2, 5, 3, 1, 4, 7, 8]);
-  });
-
-  it('complex64 rank-1', async () => {
-    const a = tf.complex([-2, -1, 0, 1, 2], [1, 2, 3, 0, -1]);
-    const result = tf.abs(a);
-    expectArraysClose(await result.data(), [
-      Math.sqrt(-2 * -2 + 1 * 1), Math.sqrt(-1 * -1 + 2 * 2),
-      Math.sqrt(0 * 0 + 3 * 3), Math.sqrt(1 * 1 + 0 * 0),
-      Math.sqrt(2 * 2 + -1 * -1)
-    ]);
-    expect(result.shape).toEqual([5]);
-  });
-
-  it('complex64 rank-2', async () => {
-    const a = tf.complex([[-3, -2, -1], [0, 1, 2]], [[4, 1, 2], [3, 0, -1]]);
-    const result = tf.abs(a);
-    expectArraysClose(await result.data(), [
-      Math.sqrt(-3 * -3 + 4 * 4), Math.sqrt(-2 * -2 + 1 * 1),
-      Math.sqrt(-1 * -1 + 2 * 2), Math.sqrt(0 * 0 + 3 * 3),
-      Math.sqrt(1 * 1 + 0 * 0), Math.sqrt(2 * 2 + -1 * -1)
-    ]);
-    expect(result.shape).toEqual([2, 3]);
-  });
-
-  it('complex64 rank-3', async () => {
-    const a = tf.complex(
-        [[[-3, -2], [-1, 0]], [[1, 2], [3, 4]]],
-        [[[4, 1], [2, 3]], [[0, -1], [-3, -4]]]);
-    const result = tf.abs(a);
-    expectArraysClose(await result.data(), [
-      Math.sqrt(-3 * -3 + 4 * 4), Math.sqrt(-2 * -2 + 1 * 1),
-      Math.sqrt(-1 * -1 + 2 * 2), Math.sqrt(0 * 0 + 3 * 3),
-      Math.sqrt(1 * 1 + 0 * 0), Math.sqrt(2 * 2 + -1 * -1),
-      Math.sqrt(3 * 3 + -3 * -3), Math.sqrt(4 * 4 + -4 * -4)
-    ]);
-    expect(result.shape).toEqual([2, 2, 2]);
-  });
-
-  it('is underflow-safe for complex64', async () => {
-    const floatBits = tf.backend().floatPrecision();
-    let small;
-    switch (floatBits) {
-      case 32:
-        small = 1e-30;
-        break;
-      case 16:
-        small = 1e-4;
-        break;
-      default:
-        throw new Error(`Test not implemented for ENV.engine.floatPrecision()=${
-            floatBits}.`);
-    }
-
-    const a = tf.complex([small, 0, small, 0], [small, small, 0, 0]);
-    const result = tf.abs(a);
-    expectArraysClose(
-        await result.data(),
-        [
-          Math.hypot(small, small), Math.hypot(0, small), Math.hypot(small, 0),
-          Math.hypot(0, 0)
-        ],
-        /*tolerance=*/small / 100);
-    expect(result.shape).toEqual([4]);
-  });
-
-  it('propagates NaNs', async () => {
-    const a = tf.tensor1d([1, -2, 0, 3, -0.1, NaN]);
-    const result = tf.abs(a);
-    expectArraysClose(await result.data(), [1, 2, 0, 3, 0.1, NaN]);
-  });
-
-  it('gradients: Scalar', async () => {
-    const a = tf.scalar(4);
-    const dy = tf.scalar(8);
-
-    const da = tf.grad(a => tf.abs(a))(a, dy);
-
-    expect(da.shape).toEqual(a.shape);
-    expect(da.dtype).toEqual('float32');
-    expectArraysClose(await da.data(), [8 * 1]);
-  });
-
-  it('gradient with clones', () => {
-    const a = tf.scalar(4);
-    const dy = tf.scalar(8);
-
-    const da = tf.grad(a => a.clone().abs().clone())(a, dy);
-
-    expect(da.shape).toEqual(a.shape);
-    expect(da.dtype).toEqual('float32');
-  });
-
-  it('gradients: Tensor1D', async () => {
-    const a = tf.tensor1d([1, 2, -3, 5]);
-    const dy = tf.tensor1d([1, 2, 3, 4]);
-
-    const da = tf.grad(a => tf.abs(a))(a, dy);
-
-    expect(da.shape).toEqual(a.shape);
-    expect(da.dtype).toEqual('float32');
-    expectArraysClose(await da.data(), [1 * 1, 2 * 1, 3 * -1, 4 * 1]);
-  });
-
-  it('gradients: Tensor2D', async () => {
-    const a = tf.tensor2d([3, -1, -2, 3], [2, 2]);
-    const dy = tf.tensor2d([1, 2, 3, 4], [2, 2]);
-
-    const da = tf.grad(a => tf.abs(a))(a, dy);
-
-    expect(da.shape).toEqual(a.shape);
-    expect(da.dtype).toEqual('float32');
-    expectArraysClose(await da.data(), [1 * 1, 2 * -1, 3 * -1, 4 * 1]);
-  });
-
-  it('throws when passed a non-tensor', () => {
-    expect(() => tf.abs({} as tf.Tensor))
-        .toThrowError(/Argument 'x' passed to 'abs' must be a Tensor/);
-  });
-
-  it('accepts a tensor-like object', async () => {
-    const result = tf.abs([1, -2, 0, 3, -0.1]);
-    expectArraysClose(await result.data(), [1, 2, 0, 3, 0.1]);
-  });
-
-  it('throws for string tensor', () => {
-    expect(() => tf.abs('q'))
-        .toThrowError(/Argument 'x' passed to 'abs' must be numeric/);
-  });
-});
+import {expectArraysClose} from '../test_util';
 
 describeWithFlags('step', ALL_ENVS, () => {
   it('with 1d tensor', async () => {
@@ -1490,911 +1342,6 @@ describeWithFlags('expm1', ALL_ENVS, () => {
   });
 });
 
-describeWithFlags('sin', ALL_ENVS, () => {
-  it('basic', async () => {
-    const values = [1, -3, 2, 7, -4];
-    const a = tf.tensor1d(values);
-    const result = tf.sin(a);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = Math.sin(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('propagates NaNs', async () => {
-    const a = tf.tensor1d([4, NaN, 0]);
-    const res = tf.sin(a);
-    expectArraysClose(await res.data(), [Math.sin(4), NaN, Math.sin(0)]);
-  });
-
-  it('gradients: Scalar', async () => {
-    const a = tf.scalar(5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.sin(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [8 * Math.cos(5)]);
-  });
-
-  it('gradient with clones', async () => {
-    const a = tf.scalar(5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.sin(a.clone()).clone())(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [8 * Math.cos(5)]);
-  });
-
-  it('gradients: Tensor1D', async () => {
-    const a = tf.tensor1d([-1, 2, 3, -5]);
-    const dy = tf.tensor1d([1, 2, 3, 4]);
-
-    const gradients = tf.grad(a => tf.sin(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(
-        await gradients.data(),
-        [1 * Math.cos(-1), 2 * Math.cos(2), 3 * Math.cos(3), 4 * Math.cos(-5)],
-        1e-1);
-  });
-
-  it('gradients: Tensor2D', async () => {
-    const a = tf.tensor2d([-3, 1, 2, 3], [2, 2]);
-    const dy = tf.tensor2d([1, 2, 3, 4], [2, 2]);
-
-    const gradients = tf.grad(a => tf.sin(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(
-        await gradients.data(),
-        [1 * Math.cos(-3), 2 * Math.cos(1), 3 * Math.cos(2), 4 * Math.cos(3)],
-        1e-1);
-  });
-
-  it('throws when passed a non-tensor', () => {
-    expect(() => tf.sin({} as tf.Tensor))
-        .toThrowError(/Argument 'x' passed to 'sin' must be a Tensor/);
-  });
-
-  it('accepts a tensor-like object', async () => {
-    const values = [1, -3, 2, 7, -4];
-    const result = tf.sin(values);
-
-    const expected = [];
-    for (let i = 0; i < values.length; i++) {
-      expected[i] = Math.sin(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('throws for string tensor', () => {
-    expect(() => tf.sin('q'))
-        .toThrowError(/Argument 'x' passed to 'sin' must be numeric/);
-  });
-});
-
-describeWithFlags('cos', ALL_ENVS, () => {
-  it('basic', async () => {
-    const values = [1, -3, 2, 7, -4];
-    const a = tf.tensor1d(values);
-    const result = tf.cos(a);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = Math.cos(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('propagates NaNs', async () => {
-    const a = tf.tensor1d([4, NaN, 0]);
-    const res = tf.cos(a);
-    expectArraysClose(await res.data(), [Math.cos(4), NaN, Math.cos(0)]);
-  });
-
-  it('gradients: Scalar', async () => {
-    const a = tf.scalar(5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.cos(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [8 * Math.sin(5) * -1]);
-  });
-
-  it('gradient with clones', async () => {
-    const a = tf.scalar(5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.cos(a.clone()).clone())(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [8 * Math.sin(5) * -1]);
-  });
-
-  it('gradients: Tensor1D', async () => {
-    const a = tf.tensor1d([-1, 2, 3, -5]);
-    const dy = tf.tensor1d([1, 2, 3, 4]);
-
-    const gradients = tf.grad(a => tf.cos(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(
-        await gradients.data(),
-        [
-          1 * Math.sin(-1) * -1, 2 * Math.sin(2) * -1, 3 * Math.sin(3) * -1,
-          4 * Math.sin(-5) * -1
-        ],
-        1e-1);
-  });
-
-  it('gradients: Tensor2D', async () => {
-    const a = tf.tensor2d([-3, 1, 2, 3], [2, 2]);
-    const dy = tf.tensor2d([1, 2, 3, 4], [2, 2]);
-
-    const gradients = tf.grad(a => tf.cos(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(
-        await gradients.data(),
-        [
-          1 * Math.sin(-3) * -1, 2 * Math.sin(1) * -1, 3 * Math.sin(2) * -1,
-          4 * Math.sin(3) * -1
-        ],
-        1e-1);
-  });
-
-  it('throws when passed a non-tensor', () => {
-    expect(() => tf.cos({} as tf.Tensor))
-        .toThrowError(/Argument 'x' passed to 'cos' must be a Tensor/);
-  });
-
-  it('accepts a tensor-like object', async () => {
-    const values = [1, -3, 2, 7, -4];
-    const result = tf.cos(values);
-
-    const expected = [];
-    for (let i = 0; i < values.length; i++) {
-      expected[i] = Math.cos(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('throws for string tensor', () => {
-    expect(() => tf.cos('q'))
-        .toThrowError(/Argument 'x' passed to 'cos' must be numeric/);
-  });
-});
-
-describeWithFlags('tan', ALL_ENVS, () => {
-  it('basic', async () => {
-    const values = [1, -3, 2, 7, -4];
-    const a = tf.tensor1d(values);
-    const result = tf.tan(a);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = Math.tan(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('propagates NaNs', async () => {
-    const a = tf.tensor1d([4, NaN, 0]);
-    const res = tf.tan(a);
-    expectArraysClose(await res.data(), [Math.tan(4), NaN, Math.tan(0)]);
-  });
-
-  it('gradients: Scalar', async () => {
-    const a = tf.scalar(0.5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.tan(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(
-        await gradients.data(), [8 / (Math.cos(0.5) * Math.cos(0.5))]);
-  });
-
-  it('gradient with clones', async () => {
-    const a = tf.scalar(0.5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.tan(a.clone()).clone())(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(
-        await gradients.data(), [8 / (Math.cos(0.5) * Math.cos(0.5))]);
-  });
-
-  it('gradients: Tensor1D', async () => {
-    const aValues = [-1, 2, 3, -5];
-    const dyValues = [1, 2, 3, 4];
-    const a = tf.tensor1d(aValues);
-    const dy = tf.tensor1d(dyValues);
-
-    const gradients = tf.grad(a => tf.tan(a))(a, dy);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = dyValues[i] / (Math.cos(aValues[i]) * Math.cos(aValues[i]));
-    }
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    // The grad(tan(x)) which relies on 1/cos(x) is less precise on Windows.
-    expectArraysClose(await gradients.data(), expected, TEST_EPSILON_FLOAT16);
-  });
-
-  it('gradients: Tensor2D', async () => {
-    const aValues = [-3, 1, 2, 3];
-    const dyValues = [1, 2, 3, 4];
-    const a = tf.tensor2d(aValues, [2, 2]);
-    const dy = tf.tensor2d(dyValues, [2, 2]);
-
-    const gradients = tf.grad(a => tf.tan(a))(a, dy);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = dyValues[i] / (Math.cos(aValues[i]) * Math.cos(aValues[i]));
-    }
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), expected);
-  });
-
-  it('throws when passed a non-tensor', () => {
-    expect(() => tf.tan({} as tf.Tensor))
-        .toThrowError(/Argument 'x' passed to 'tan' must be a Tensor/);
-  });
-
-  it('accepts a tensor-like object', async () => {
-    const values = [1, -3, 2, 7, -4];
-    const result = tf.tan(values);
-
-    const expected = [];
-    for (let i = 0; i < values.length; i++) {
-      expected[i] = Math.tan(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('throws for string tensor', () => {
-    expect(() => tf.tan('q'))
-        .toThrowError(/Argument 'x' passed to 'tan' must be numeric/);
-  });
-});
-
-describeWithFlags('asin', ALL_ENVS, () => {
-  it('basic', async () => {
-    const values = [.1, -3, 2, 7, -4];
-    const a = tf.tensor1d(values);
-    const result = tf.asin(a);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = Math.asin(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('propagates NaNs', async () => {
-    const a = tf.tensor1d([4, NaN, 0]);
-    const res = tf.asin(a);
-    expectArraysClose(await res.data(), [Math.asin(4), NaN, Math.asin(0)]);
-  });
-
-  it('gradients: Scalar', async () => {
-    const a = tf.scalar(0.5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.asin(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [8 / Math.sqrt(1 - (0.5 * 0.5))]);
-  });
-
-  it('gradient with clones', async () => {
-    const a = tf.scalar(0.5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.asin(a.clone()).clone())(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [8 / Math.sqrt(1 - (0.5 * 0.5))]);
-  });
-
-  it('gradients: Tensor1D', async () => {
-    const aValues = [-0.1, 0.2, 0.3, -0.5];
-    const dyValues = [1, 2, 3, 4];
-    const a = tf.tensor1d(aValues);
-    const dy = tf.tensor1d(dyValues);
-
-    const gradients = tf.grad(a => tf.asin(a))(a, dy);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = dyValues[i] / Math.sqrt(1 - (aValues[i] * aValues[i]));
-    }
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), expected);
-  });
-
-  it('gradients: Tensor2D', async () => {
-    const aValues = [-0.3, 0.1, 0.2, 0.3];
-    const dyValues = [1, 2, 3, 4];
-    const a = tf.tensor2d(aValues, [2, 2]);
-    const dy = tf.tensor2d(dyValues, [2, 2]);
-
-    const gradients = tf.grad(a => tf.asin(a))(a, dy);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = dyValues[i] / Math.sqrt(1 - (aValues[i] * aValues[i]));
-    }
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), expected);
-  });
-
-  it('throws when passed a non-tensor', () => {
-    expect(() => tf.asin({} as tf.Tensor))
-        .toThrowError(/Argument 'x' passed to 'asin' must be a Tensor/);
-  });
-
-  it('accepts a tensor-like object', async () => {
-    const values = [.1, -3, 2, 7, -4];
-    const result = tf.asin(values);
-
-    const expected = [];
-    for (let i = 0; i < values.length; i++) {
-      expected[i] = Math.asin(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('throws for string tensor', () => {
-    expect(() => tf.asin('q'))
-        .toThrowError(/Argument 'x' passed to 'asin' must be numeric/);
-  });
-});
-
-describeWithFlags('acos', ALL_ENVS, () => {
-  it('basic', async () => {
-    const values = [.1, -3, 2, 7, -4];
-    const a = tf.tensor1d(values);
-    const result = tf.acos(a);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = Math.acos(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('propagates NaNs', async () => {
-    const a = tf.tensor1d([4, NaN, 0]);
-    const res = tf.acos(a);
-    expectArraysClose(await res.data(), [Math.acos(4), NaN, Math.acos(0)]);
-  });
-
-  it('gradients: Scalar', async () => {
-    const a = tf.scalar(0.5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.acos(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(
-        await gradients.data(), [(-1 * 8) / Math.sqrt(1 - (0.5 * 0.5))]);
-  });
-
-  it('gradient with clones', async () => {
-    const a = tf.scalar(0.5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.acos(a.clone()).clone())(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(
-        await gradients.data(), [(-1 * 8) / Math.sqrt(1 - (0.5 * 0.5))]);
-  });
-
-  it('gradients: Tensor1D', async () => {
-    const aValues = [-0.1, 0.2, 0.3, -0.5];
-    const dyValues = [1, 2, 3, 4];
-    const a = tf.tensor1d(aValues);
-    const dy = tf.tensor1d(dyValues);
-
-    const gradients = tf.grad(a => tf.acos(a))(a, dy);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] =
-          (-1 * dyValues[i]) / Math.sqrt(1 - (aValues[i] * aValues[i]));
-    }
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), expected);
-  });
-
-  it('gradients: Tensor2D', async () => {
-    const aValues = [-0.3, 0.1, 0.2, 0.3];
-    const dyValues = [1, 2, 3, 4];
-    const a = tf.tensor2d(aValues, [2, 2]);
-    const dy = tf.tensor2d(dyValues, [2, 2]);
-
-    const gradients = tf.grad(a => tf.acos(a))(a, dy);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] =
-          (-1 * dyValues[i]) / Math.sqrt(1 - (aValues[i] * aValues[i]));
-    }
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), expected);
-  });
-
-  it('throws when passed a non-tensor', () => {
-    expect(() => tf.acos({} as tf.Tensor))
-        .toThrowError(/Argument 'x' passed to 'acos' must be a Tensor/);
-  });
-
-  it('accepts a tensor-like object', async () => {
-    const values = [.1, -3, 2, 7, -4];
-    const result = tf.acos(values);
-
-    const expected = [];
-    for (let i = 0; i < values.length; i++) {
-      expected[i] = Math.acos(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('throws for string tensor', () => {
-    expect(() => tf.acos('q'))
-        .toThrowError(/Argument 'x' passed to 'acos' must be numeric/);
-  });
-});
-
-describeWithFlags('atan', ALL_ENVS, () => {
-  it('basic', async () => {
-    const values = [1, -3, 2, 7, -4];
-    const a = tf.tensor1d(values);
-    const result = tf.atan(a);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = Math.atan(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('6D atan', async () => {
-    const a = tf.range(1, 65).reshape([2, 2, 2, 2, 2, 2]);
-    const result = tf.atan(a);
-
-    const expected = [];
-    for (let i = 1; i < 65; ++i) {
-      expected[i - 1] = Math.atan(i);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('propagates NaNs', async () => {
-    const a = tf.tensor1d([4, NaN, 0]);
-    const res = tf.atan(a);
-    expectArraysClose(await res.data(), [Math.atan(4), NaN, Math.atan(0)]);
-  });
-
-  it('gradients: Scalar', async () => {
-    const a = tf.scalar(0.5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.atan(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [8 / (1 + (0.5 * 0.5))]);
-  });
-
-  it('gradient with clones', async () => {
-    const a = tf.scalar(0.5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.atan(a.clone()).clone())(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [8 / (1 + (0.5 * 0.5))]);
-  });
-
-  it('gradients: Tensor1D', async () => {
-    const aValues = [-0.1, 0.2, 0.3, -0.5];
-    const dyValues = [1, 2, 3, 4];
-    const a = tf.tensor1d(aValues);
-    const dy = tf.tensor1d(dyValues);
-
-    const gradients = tf.grad(a => tf.atan(a))(a, dy);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = dyValues[i] / (1 + (aValues[i] * aValues[i]));
-    }
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), expected);
-  });
-
-  it('gradients: Tensor2D', async () => {
-    const aValues = [-0.3, 0.1, 0.2, 0.3];
-    const dyValues = [1, 2, 3, 4];
-    const a = tf.tensor2d(aValues, [2, 2]);
-    const dy = tf.tensor2d(dyValues, [2, 2]);
-
-    const gradients = tf.grad(a => tf.atan(a))(a, dy);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = dyValues[i] / (1 + (aValues[i] * aValues[i]));
-    }
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), expected);
-  });
-
-  it('throws when passed a non-tensor', () => {
-    expect(() => tf.atan({} as tf.Tensor))
-        .toThrowError(/Argument 'x' passed to 'atan' must be a Tensor/);
-  });
-
-  it('accepts a tensor-like object', async () => {
-    const values = [1, -3, 2, 7, -4];
-    const result = tf.atan(values);
-
-    const expected = [];
-    for (let i = 0; i < values.length; i++) {
-      expected[i] = Math.atan(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('throws for string tensor', () => {
-    expect(() => tf.atan('q'))
-        .toThrowError(/Argument 'x' passed to 'atan' must be numeric/);
-  });
-});
-
-describeWithFlags('sinh', ALL_ENVS, () => {
-  it('basic', async () => {
-    const values = [1, -3, 2, -1, -4];
-    const a = tf.tensor1d(values);
-    const result = tf.sinh(a);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = Math.sinh(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('propagates NaNs', async () => {
-    const a = tf.tensor1d([4, NaN, 0]);
-    const res = tf.sinh(a);
-    expectArraysClose(await res.data(), [Math.sinh(4), NaN, Math.sinh(0)]);
-  });
-
-  it('gradients: Scalar', async () => {
-    const a = tf.scalar(0.5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.sinh(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [8 * Math.cosh(0.5)]);
-  });
-
-  it('gradient with clones', async () => {
-    const a = tf.scalar(0.5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.sinh(a.clone()).clone())(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [8 * Math.cosh(0.5)]);
-  });
-
-  it('gradients: Tensor1D', async () => {
-    const aValues = [-1, 2, 3, -5];
-    const dyValues = [1, 2, 3, 4];
-    const a = tf.tensor1d(aValues);
-    const dy = tf.tensor1d(dyValues);
-
-    const gradients = tf.grad(a => tf.sinh(a))(a, dy);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = dyValues[i] * Math.cosh(aValues[i]);
-    }
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), expected);
-  });
-
-  it('gradients: Tensor2D', async () => {
-    const aValues = [-3, 1, 2, 3];
-    const dyValues = [1, 2, 3, 4];
-    const a = tf.tensor2d(aValues, [2, 2]);
-    const dy = tf.tensor2d(dyValues, [2, 2]);
-
-    const gradients = tf.grad(a => tf.sinh(a))(a, dy);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = dyValues[i] * Math.cosh(aValues[i]);
-    }
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), expected);
-  });
-
-  it('throws when passed a non-tensor', () => {
-    expect(() => tf.sinh({} as tf.Tensor))
-        .toThrowError(/Argument 'x' passed to 'sinh' must be a Tensor/);
-  });
-
-  it('accepts a tensor-like object', async () => {
-    const values = [1, -3, 2, -1, -4];
-    const result = tf.sinh(values);
-
-    const expected = [];
-    for (let i = 0; i < values.length; i++) {
-      expected[i] = Math.sinh(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('throws for string tensor', () => {
-    expect(() => tf.sinh('q'))
-        .toThrowError(/Argument 'x' passed to 'sinh' must be numeric/);
-  });
-});
-
-describeWithFlags('cosh', ALL_ENVS, () => {
-  it('basic', async () => {
-    const values = [1, -3, 2, -1, -4];
-    const a = tf.tensor1d(values);
-    const result = tf.cosh(a);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = Math.cosh(values[i]);
-    }
-
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('propagates NaNs', async () => {
-    const a = tf.tensor1d([4, NaN, 0]);
-    const res = tf.cosh(a);
-    expectArraysClose(await res.data(), [Math.cosh(4), NaN, Math.cosh(0)]);
-  });
-
-  it('gradients: Scalar', async () => {
-    const a = tf.scalar(0.5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.cosh(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [8 * Math.sinh(0.5)]);
-  });
-
-  it('gradient with clones', async () => {
-    const a = tf.scalar(0.5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.cosh(a.clone()).clone())(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [8 * Math.sinh(0.5)]);
-  });
-
-  it('gradients: Tensor1D', async () => {
-    const aValues = [-1, 2, 3, -5];
-    const dyValues = [1, 2, 3, 4];
-    const a = tf.tensor1d(aValues);
-    const dy = tf.tensor1d(dyValues);
-
-    const gradients = tf.grad(a => tf.cosh(a))(a, dy);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = dyValues[i] * Math.sinh(aValues[i]);
-    }
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), expected);
-  });
-
-  it('gradients: Tensor2D', async () => {
-    const aValues = [-3, 1, 2, 3];
-    const dyValues = [1, 2, 3, 4];
-    const a = tf.tensor2d(aValues, [2, 2]);
-    const dy = tf.tensor2d(dyValues, [2, 2]);
-
-    const gradients = tf.grad(a => tf.cosh(a))(a, dy);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = dyValues[i] * Math.sinh(aValues[i]);
-    }
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), expected);
-  });
-
-  it('throws when passed a non-tensor', () => {
-    expect(() => tf.cosh({} as tf.Tensor))
-        .toThrowError(/Argument 'x' passed to 'cosh' must be a Tensor/);
-  });
-
-  it('accepts a tensor-like object', async () => {
-    const values = [1, -3, 2, -1, -4];
-    const result = tf.cosh(values);
-
-    const expected = [];
-    for (let i = 0; i < values.length; i++) {
-      expected[i] = Math.cosh(values[i]);
-    }
-
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('throws for string tensor', () => {
-    expect(() => tf.cosh('q'))
-        .toThrowError(/Argument 'x' passed to 'cosh' must be numeric/);
-  });
-});
-
-describeWithFlags('tanh', ALL_ENVS, () => {
-  it('basic', async () => {
-    const values = [1, -3, 2, 7, -4];
-    const a = tf.tensor1d(values);
-    const result = tf.tanh(a);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = util.tanh(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('propagates NaNs', async () => {
-    const a = tf.tensor1d([4, NaN, 0]);
-    const res = tf.tanh(a);
-    expectArraysClose(await res.data(), [util.tanh(4), NaN, util.tanh(0)]);
-  });
-
-  it('gradients: Scalar', async () => {
-    const a = tf.scalar(0.5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.tanh(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(
-        await gradients.data(), [8 * (1 - (Math.tanh(0.5) * Math.tanh(0.5)))]);
-  });
-
-  it('gradient with clones', async () => {
-    const a = tf.scalar(0.5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.tanh(a.clone()).clone())(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(
-        await gradients.data(), [8 * (1 - (Math.tanh(0.5) * Math.tanh(0.5)))]);
-  });
-
-  it('gradients: Tensor1D', async () => {
-    const aValues = [-1, 2, 3, -5];
-    const dyValues = [1, 2, 3, 4];
-    const a = tf.tensor1d(aValues);
-    const dy = tf.tensor1d(dyValues);
-
-    const gradients = tf.grad(a => tf.tanh(a))(a, dy);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] =
-          dyValues[i] * (1 - (Math.tanh(aValues[i]) * Math.tanh(aValues[i])));
-    }
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), expected);
-  });
-
-  it('gradients: Tensor2D', async () => {
-    const aValues = [-3, 1, 2, 3];
-    const dyValues = [1, 2, 3, 4];
-    const a = tf.tensor2d(aValues, [2, 2]);
-    const dy = tf.tensor2d(dyValues, [2, 2]);
-
-    const gradients = tf.grad(a => tf.tanh(a))(a, dy);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] =
-          dyValues[i] * (1 - (Math.tanh(aValues[i]) * Math.tanh(aValues[i])));
-    }
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), expected);
-  });
-
-  it('throws when passed a non-tensor', () => {
-    expect(() => tf.tanh({} as tf.Tensor))
-        .toThrowError(/Argument 'x' passed to 'tanh' must be a Tensor/);
-  });
-
-  it('accepts a tensor-like object', async () => {
-    const values = [1, -3, 2, 7, -4];
-    const result = tf.tanh(values);
-
-    const expected = [];
-    for (let i = 0; i < values.length; i++) {
-      expected[i] = util.tanh(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('throws for string tensor', () => {
-    expect(() => tf.tanh('q'))
-        .toThrowError(/Argument 'x' passed to 'tanh' must be numeric/);
-  });
-});
-
 describeWithFlags('clip', ALL_ENVS, () => {
   it('basic', async () => {
     const a = tf.tensor1d([3, -1, 0, 100, -7, 2]);
@@ -2584,373 +1531,6 @@ describeWithFlags('round', ALL_ENVS, () => {
   it('throws for string tensor', () => {
     expect(() => tf.round('q'))
         .toThrowError(/Argument 'x' passed to 'round' must be numeric/);
-  });
-});
-
-describeWithFlags('asinh', ALL_ENVS, () => {
-  it('basic', async () => {
-    const values = [1, -3, 2, 7, -4];
-    const a = tf.tensor1d(values);
-    const result = tf.asinh(a);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = Math.asinh(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('scalar', async () => {
-    const a = tf.scalar(1);
-    const result = tf.asinh(a);
-
-    const expected = [Math.asinh(1)];
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('tensor2D', async () => {
-    const values = [1, -3, 2, 7];
-    const a = tf.tensor2d(values, [2, 2]);
-    const result = tf.asinh(a);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = Math.asinh(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('propagates NaNs', async () => {
-    const a = tf.tensor1d([4, NaN, 0]);
-    const res = tf.asinh(a);
-    expectArraysClose(await res.data(), [Math.asinh(4), NaN, Math.asinh(0)]);
-  });
-
-  it('gradients: Scalar', async () => {
-    const a = tf.scalar(0.5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.asinh(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [8 / Math.sqrt(1.0 + 0.5 * 0.5)]);
-  });
-
-  it('gradient with clones', async () => {
-    const a = tf.scalar(0.5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.asinh(a.clone()).clone())(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [8 / Math.sqrt(1.0 + 0.5 * 0.5)]);
-  });
-
-  it('gradients: Tensor1D', async () => {
-    const aValues = [-1, 2, 3, -5];
-    const dyValues = [1, 2, 3, 4];
-    const a = tf.tensor1d(aValues);
-    const dy = tf.tensor1d(dyValues);
-
-    const gradients = tf.grad(a => tf.asinh(a))(a, dy);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = dyValues[i] / Math.sqrt(1 + aValues[i] * aValues[i]);
-    }
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), expected);
-  });
-
-  it('gradients: Tensor2D', async () => {
-    const aValues = [-3, 1, 2, 3];
-    const dyValues = [1, 2, 3, 4];
-    const a = tf.tensor2d(aValues, [2, 2]);
-    const dy = tf.tensor2d(dyValues, [2, 2]);
-
-    const gradients = tf.grad(a => tf.asinh(a))(a, dy);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = dyValues[i] / Math.sqrt(1 + aValues[i] * aValues[i]);
-    }
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), expected);
-  });
-
-  it('throws when passed a non-tensor', () => {
-    expect(() => tf.asinh({} as tf.Tensor))
-        .toThrowError(/Argument 'x' passed to 'asinh' must be a Tensor/);
-  });
-
-  it('accepts a tensor-like object', async () => {
-    const values = [1, -3, 2, 7, -4];
-    const result = tf.asinh(values);
-
-    const expected = [];
-    for (let i = 0; i < values.length; i++) {
-      expected[i] = Math.asinh(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('throws for string tensor', () => {
-    expect(() => tf.asinh('q'))
-        .toThrowError(/Argument 'x' passed to 'asinh' must be numeric/);
-  });
-});
-
-describeWithFlags('acosh', ALL_ENVS, () => {
-  it('basic', async () => {
-    const values = [2, 3, 4, 5, 6];
-    const a = tf.tensor1d(values);
-    const result = tf.acosh(a);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = Math.acosh(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('scalar', async () => {
-    const value = 2;
-    const a = tf.scalar(value);
-    const result = tf.acosh(a);
-
-    const expected = [Math.acosh(value)];
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('tensor2d', async () => {
-    const values = [2, 3, 4, 5];
-    const a = tf.tensor2d(values, [2, 2]);
-    const result = tf.acosh(a);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = Math.acosh(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('propagates NaNs', async () => {
-    const a = tf.tensor1d([4, NaN, 2]);
-    const res = tf.acosh(a);
-    expectArraysClose(await res.data(), [Math.acosh(4), NaN, Math.acosh(2)]);
-  });
-
-  it('NaN outside function domain', async () => {
-    const a = tf.tensor1d([4, -1, 2]);
-    const res = tf.acosh(a);
-    expectArraysClose(await res.data(), [Math.acosh(4), NaN, Math.acosh(2)]);
-  });
-
-  it('gradients: Scalar', async () => {
-    const a = tf.scalar(1.5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.acosh(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(
-        await gradients.data(), [8.0 / Math.sqrt(1.5 * 1.5 - 1.0)]);
-  });
-
-  it('gradient with clones', async () => {
-    const a = tf.scalar(1.5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.acosh(a.clone()).clone())(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(
-        await gradients.data(), [8.0 / Math.sqrt(1.5 * 1.5 - 1.0)]);
-  });
-
-  it('gradients: Tensor1D', async () => {
-    const aValues = [2, 3, 5, 10];
-    const dyValues = [1, 2, 3, 4];
-    const a = tf.tensor1d(aValues);
-    const dy = tf.tensor1d(dyValues);
-
-    const gradients = tf.grad(a => tf.acosh(a))(a, dy);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = dyValues[i] / Math.sqrt(Math.pow(aValues[i], 2) - 1.0);
-    }
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), expected);
-  });
-
-  it('gradients: Tensor2D', async () => {
-    const aValues = [2, 3, 5, 7];
-    const dyValues = [1, 2, 3, 4];
-    const a = tf.tensor2d(aValues, [2, 2]);
-    const dy = tf.tensor2d(dyValues, [2, 2]);
-
-    const gradients = tf.grad(a => tf.acosh(a))(a, dy);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = dyValues[i] / Math.sqrt(Math.pow(aValues[i], 2) - 1.0);
-    }
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), expected);
-  });
-
-  it('throws when passed a non-tensor', () => {
-    expect(() => tf.acosh({} as tf.Tensor))
-        .toThrowError(/Argument 'x' passed to 'acosh' must be a Tensor/);
-  });
-
-  it('accepts a tensor-like object', async () => {
-    const values = [2, 3, 4, 5, 6];
-    const result = tf.acosh(values);
-
-    const expected = [];
-    for (let i = 0; i < values.length; i++) {
-      expected[i] = Math.acosh(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('throws for string tensor', () => {
-    expect(() => tf.acosh('q'))
-        .toThrowError(/Argument 'x' passed to 'acosh' must be numeric/);
-  });
-});
-
-describeWithFlags('atanh', ALL_ENVS, () => {
-  it('basic', async () => {
-    const values = [-0.25, 0.25, 0.5, .75, -0.4];
-    const a = tf.tensor1d(values);
-    const result = tf.atanh(a);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = Math.atanh(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('scalar', async () => {
-    const value = 0.2;
-    const a = tf.scalar(value);
-    const result = tf.atanh(a);
-
-    const expected = [Math.atanh(value)];
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('tensor2d', async () => {
-    const values = [0.2, 0.3, 0.4, 0.5];
-    const a = tf.tensor2d(values, [2, 2]);
-    const result = tf.atanh(a);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = Math.atanh(values[i]);
-    }
-    expectArraysClose(await result.data(), expected);
-  });
-
-  it('propagates NaNs', async () => {
-    const a = tf.tensor1d([0.5, NaN, 0]);
-    const res = tf.atanh(a);
-    expectArraysClose(await res.data(), [Math.atanh(0.5), NaN, Math.atanh(0)]);
-  });
-
-  it('NaN outside function domain', async () => {
-    const a = tf.tensor1d([-2, 0, 2]);
-    const res = tf.atanh(a);
-    expectArraysClose(await res.data(), [NaN, Math.atanh(0), NaN]);
-  });
-
-  it('gradients: Scalar', async () => {
-    const a = tf.scalar(0.5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.atanh(a))(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [8 / (1 - 0.5 * 0.5)]);
-  });
-
-  it('gradient with clones', async () => {
-    const a = tf.scalar(0.5);
-    const dy = tf.scalar(8);
-
-    const gradients = tf.grad(a => tf.atanh(a.clone()).clone())(a, dy);
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), [8 / (1 - 0.5 * 0.5)]);
-  });
-
-  it('gradients: Tensor1D', async () => {
-    const aValues = [-0.1, 0.2, 0.3, -0.5];
-    const dyValues = [1, 2, 3, 4];
-    const a = tf.tensor1d(aValues);
-    const dy = tf.tensor1d(dyValues);
-
-    const gradients = tf.grad(a => tf.atanh(a))(a, dy);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = dyValues[i] / (1 - Math.pow(aValues[i], 2));
-    }
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), expected);
-  });
-
-  it('gradients: Tensor2D', async () => {
-    const aValues = [-0.3, 0.1, 0.2, 0.3];
-    const dyValues = [1, 2, 3, 4];
-    const a = tf.tensor2d(aValues, [2, 2]);
-    const dy = tf.tensor2d(dyValues, [2, 2]);
-
-    const gradients = tf.grad(a => tf.atanh(a))(a, dy);
-
-    const expected = [];
-    for (let i = 0; i < a.size; i++) {
-      expected[i] = dyValues[i] / (1 - Math.pow(aValues[i], 2));
-    }
-
-    expect(gradients.shape).toEqual(a.shape);
-    expect(gradients.dtype).toEqual('float32');
-    expectArraysClose(await gradients.data(), expected);
-  });
-
-  it('throws when passed a non-tensor', () => {
-    expect(() => tf.atanh({} as tf.Tensor))
-        .toThrowError(/Argument 'x' passed to 'atanh' must be a Tensor/);
-  });
-
-  it('accepts a tensor-like object', async () => {
-    const result = tf.atanh(0.2);
-    expectArraysClose(await result.data(), [Math.atanh(0.2)]);
-  });
-
-  it('throws for string tensor', () => {
-    expect(() => tf.atanh('q'))
-        .toThrowError(/Argument 'x' passed to 'atanh' must be numeric/);
   });
 });
