+++ tfjs-core/src/register_all_gradients.ts
@@ -14,11 +14,18 @@
  * limitations under the License.
  * =============================================================================
  */
+import {absGradConfig} from './gradients/Abs_grad';
+import {acosGradConfig} from './gradients/Acos_grad';
+import {acoshGradConfig} from './gradients/Acosh_grad';
 import {addGradConfig} from './gradients/Add_grad';
 import {addNGradConfig} from './gradients/AddN_grad';
 import {argMaxGradConfig} from './gradients/ArgMax_grad';
 import {argMinGradConfig} from './gradients/ArgMin_grad';
+import {asinGradConfig} from './gradients/Asin_grad';
+import {asinhGradConfig} from './gradients/Asinh_grad';
 import {atan2GradConfig} from './gradients/Atan2_grad';
+import {atanGradConfig} from './gradients/Atan_grad';
+import {atanhGradConfig} from './gradients/Atanh_grad';
 import {avgPool3DGradConfig} from './gradients/AvgPool3D_grad';
 import {avgPoolGradConfig} from './gradients/AvgPool_grad';
 import {batchMatMulGradConfig} from './gradients/BatchMatMul_grad';
@@ -29,6 +36,8 @@ import {concatGradConfig} from './gradients/Concat_grad';
 import {conv2DGradConfig} from './gradients/Conv2D_grad';
 import {conv2DBackpropInputGradConfig} from './gradients/Conv2DBackpropInput_grad';
 import {conv3DGradConfig} from './gradients/Conv3D_grad';
+import {cosGradConfig} from './gradients/Cos_grad';
+import {coshGradConfig} from './gradients/Cosh_grad';
 import {cumsumGradConfig} from './gradients/Cumsum_grad';
 import {depthwiseConv2dNativeGradConfig} from './gradients/DepthwiseConv2dNative_grad';
 import {dilation2dGradConfig} from './gradients/Dilation2D_grad';
@@ -63,12 +72,16 @@ import {reverseGradConfig} from './gradients/Reverse_grad';
 import {selectV2PoolGradConfig} from './gradients/SelectV2_grad';
 import {seluGradConfig} from './gradients/Selu_grad';
 import {signGradConfig} from './gradients/Sign_grad';
+import {sinGradConfig} from './gradients/Sin_grad';
+import {sinhGradConfig} from './gradients/Sinh_grad';
 import {spaceToBatchNDGradConfig} from './gradients/SpaceToBatchND_grad';
 import {splitVGradConfig} from './gradients/SplitV_grad';
 import {squareGradConfig} from './gradients/Square_grad';
 import {squaredDifferenceGradConfig} from './gradients/SquaredDifference_grad';
 import {subGradConfig} from './gradients/Sub_grad';
 import {sumGradConfig} from './gradients/Sum_grad';
+import {tanGradConfig} from './gradients/Tan_grad';
+import {tanhGradConfig} from './gradients/Tanh_grad';
 import {tileGradConfig} from './gradients/Tile_grad';
 import {transposeGradConfig} from './gradients/Transpose_grad';
 import {unpackGradConfig} from './gradients/Unpack_grad';
@@ -78,11 +91,18 @@ import {registerGradient} from './kernel_registry';
 
 // Export all kernel configs here so that the package can auto register them
 const gradConfigs: GradConfig[] = [
+  absGradConfig,
+  acosGradConfig,
+  acoshGradConfig,
   addGradConfig,
   addNGradConfig,
   argMaxGradConfig,
   argMinGradConfig,
+  asinGradConfig,
+  asinhGradConfig,
   atan2GradConfig,
+  atanGradConfig,
+  atanhGradConfig,
   avgPool3DGradConfig,
   avgPoolGradConfig,
   batchMatMulGradConfig,
@@ -93,6 +113,8 @@ const gradConfigs: GradConfig[] = [
   conv2DBackpropInputGradConfig,
   conv2DGradConfig,
   conv3DGradConfig,
+  cosGradConfig,
+  coshGradConfig,
   cumsumGradConfig,
   depthwiseConv2dNativeGradConfig,
   dilation2dGradConfig,
@@ -130,6 +152,8 @@ const gradConfigs: GradConfig[] = [
   selectV2PoolGradConfig,
   seluGradConfig,
   signGradConfig,
+  sinGradConfig,
+  sinhGradConfig,
   spaceToBatchNDGradConfig,
   spaceToBatchNDGradConfig,
   splitVGradConfig,
@@ -138,6 +162,8 @@ const gradConfigs: GradConfig[] = [
   squareGradConfig,
   subGradConfig,
   sumGradConfig,
+  tanGradConfig,
+  tanhGradConfig,
   tileGradConfig,
   transposeGradConfig,
   unpackGradConfig,
