+++ tfjs-core/src/tests.ts
@@ -38,7 +38,10 @@ import './io/router_registry_test';
 import './io/weights_loader_test';
 import './jasmine_util_test';
 import './kernel_registry_test';
+import './ops/abs_test';
 import './ops/absolute_difference_test';
+import './ops/acos_test';
+import './ops/acosh_test';
 import './ops/add_n_test';
 import './ops/add_test';
 import './ops/all_test';
@@ -47,6 +50,10 @@ import './ops/arg_max_test';
 import './ops/arg_min_test';
 import './ops/arithmetic_test';
 import './ops/array_ops_test';
+import './ops/asin_test';
+import './ops/asinh_test';
+import './ops/atan_test';
+import './ops/atanh_test';
 import './ops/avg_pool_3d_test';
 import './ops/avg_pool_test';
 import './ops/axis_util_test';
@@ -74,6 +81,8 @@ import './ops/conv2d_transpose_test';
 import './ops/conv3d_test';
 import './ops/conv3d_transpose_test';
 import './ops/conv_util_test';
+import './ops/cos_test';
+import './ops/cosh_test';
 import './ops/cosine_distance_test';
 import './ops/crop_and_resize_test';
 import './ops/cumsum_test';
@@ -147,6 +156,8 @@ import './ops/selu_test';
 import './ops/sigmoid_cross_entropy_test';
 import './ops/sign_test';
 import './ops/signal_ops_test';
+import './ops/sin_test';
+import './ops/sinh_test';
 import './ops/slice_test';
 import './ops/slice_util_test';
 import './ops/softmax_cross_entropy_test';
@@ -158,6 +169,8 @@ import './ops/stack_test';
 import './ops/strided_slice_test';
 import './ops/sub_test';
 import './ops/sum_test';
+import './ops/tan_test';
+import './ops/tanh_test';
 import './ops/tile_test';
 import './ops/topk_test';
 import './ops/transpose_test';
